requirements.txt
bcrypt==3.2.2
boto3==1.23.10
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
createsend==7.0.0
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-extensions==0.1.3
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
DateTime==4.4
db-dtypes==1.0.2
docutils==0.18.1
Flask==2.1.2
Flask-Caching==1.11.1
Flask-Compress==1.12
future==0.18.2
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.14.1
google-cloud-core==2.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
more-itertools==8.13.0
numpy==1.22.4
pandas==1.4.2
plotly==5.8.0
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
retrying==1.3.3
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zope.interface==5.4.0
